refactor: composants génériques pour comparaison des territoires
