--- src/client/components/PageChantier/ComparaisonTerritoires/ComparaisonTerritoires.interface.ts
@@ -1,1 +1,0 @@
-export type TypeCarte = "ta" | "meteo" | "pva";
