+++ src/client/components/_commons/ComparaisonTerritoires/PanneauCarte.tsx
@@ -1,0 +1,56 @@
+import { ReactNode } from "react";
+import { Icone } from "@/components/_commons/Icone";
+import { DeleteIcon } from "@/components/_commons/Icones/DeleteIcon";
+import { SelecteurTypeCarte } from "./SelecteurTypeCarte";
+
+type PanneauCarteProps<T extends string> = {
+  typeCarte: T;
+  options: { value: T; label: string }[];
+  estEnComparaison: boolean;
+  onChangerType: (type: T) => void;
+  onComparer: () => void;
+  onSupprimer: () => void;
+  renderCarte: (typeCarte: T) => ReactNode;
+};
+
+export const PanneauCarte = <T extends string>({
+  typeCarte,
+  options,
+  estEnComparaison,
+  onChangerType,
+  onComparer,
+  onSupprimer,
+  renderCarte,
+}: PanneauCarteProps<T>) => {
+  return (
+    <div className="flex flex-col gap-4">
+      <div className="flex flex-col justify-between gap-2">
+        <SelecteurTypeCarte
+          typeCarte={typeCarte}
+          options={options}
+          onChange={onChangerType}
+        />
+        {estEnComparaison ? (
+          <button
+            className="fr-btn fr-btn--tertiary-no-outline fr-btn--sm"
+            onClick={onSupprimer}
+            type="button"
+          >
+            <Icone className="w-4 h-4 mr-1" icone={DeleteIcon} />
+            supprimer la carte
+          </button>
+        ) : (
+          <button
+            className="fr-btn fr-btn--tertiary-no-outline fr-btn--sm"
+            onClick={onComparer}
+            type="button"
+          >
+            + comparer avec une autre carte
+          </button>
+        )}
+      </div>
+
+      {renderCarte(typeCarte)}
+    </div>
+  );
+};
