--- src/client/components/_commons/IndicateursChantier/Bloc/Détails/ComparaisonTerritoires/ComparaisonTerritoires.interface.ts
@@ -1,1 +1,0 @@
-export type TypeCarteIndicateur = "ta" | "valeurAvancement";
